databricks/requirements.txt
pyspark==3.5.0
datadog==0.49.1
python-dotenv==1.0.0
requests==2.31.0
